refactor(java_analyzer): split 131K single file into modular package
Decomposed ai-engine/agents/java_analyzer.py (131K chars, 3208 lines) into
a package with 6 independently testable submodules:

- archive_reader.py: JAR/ZIP extraction
- framework_detector.py: Forge/Fabric/Quilt detection
- feature_extractor.py: AST-based feature extraction
- embedding_bridge.py: embedding generation calls
- llm_analyzer.py: LLM complexity analysis
- tools.py: CrewAI @tool wrappers

Public API unchanged - JavaAnalyzerAgent re-exported from package.
Removed C901 suppression from pyproject.toml for this module.

Fixes #1099
