+++ ai-engine/agents/java_analyzer/__init__.py
@@ -1,0 +1,23 @@
+"""
+Java Analyzer package - modular Java mod analysis.
+
+Provides the same public API as the original java_analyzer module
+by re-exporting JavaAnalyzerAgent from java_analyzer.java_analyzer.
+"""
+
+from models.smart_assumptions import SmartAssumptionEngine
+from utils.embedding_generator import LocalEmbeddingGenerator
+
+from agents.java_analyzer.java_analyzer import JavaAnalyzerAgent
+from agents.java_analyzer.feature_extractor import JAVASSIST_AVAILABLE, TREE_SITTER_AVAILABLE
+
+JavaAnalyzer = JavaAnalyzerAgent
+
+__all__ = [
+    "JavaAnalyzerAgent",
+    "JavaAnalyzer",
+    "JAVASSIST_AVAILABLE",
+    "TREE_SITTER_AVAILABLE",
+    "SmartAssumptionEngine",
+    "LocalEmbeddingGenerator",
+]
