+++ ai-engine/pyproject.toml
@@ -89,7 +89,7 @@ ignore = [
 "agents/validation_agent.py" = ["C901"]
 "benchmarking/performance_system.py" = ["C901"]
 "crew/__init__.py" = ["N802"]  # Function name should be lowercase (factory functions RAGTasks, RAGCrew)
-"agents/java_analyzer.py" = ["N806", "C901"]  # Variable in function should be lowercase (RuntimeVisibleAnnotations)
+"agents/java_analyzer/" = ["N806"]  # Variable in function should be lowercase (RuntimeVisibleAnnotations) - split into submodules
 "tools/search_tool.py" = ["N806"]  # Variable in function should be lowercase (FallbackToolClass)
 "agents/audio_converter.py" = ["F401", "F822"]  # Optional converter_utils import, __all__ references private functions
 "agents/model_converter.py" = ["F401", "C901", "F822"]  # Optional converter_utils import, __all__ references private functions
